requirements.txt
pyautogui==0.9.53
pyserial==3.5
